renamed module: ed25519sig to ed25519msg
